# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Davomat tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/checkin:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Employee check-in (kelish)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "200":
          description: Checked in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/checkout:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Employee check-out (ketish)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutRequest"
      responses:
        "200":
          description: Checked out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /leave:
    get:
      operationId: listLeaveRequests
      tags: [leave]
      summary: List leave requests
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leave]
      summary: Create leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveRequest"
      responses:
        "201":
          description: Leave request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/{id}:
    put:
      operationId: updateLeaveRequest
      tags: [leave]
      summary: Approve/reject leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveRequest"
      responses:
        "200":
          description: Leave request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
    delete:
      operationId: deleteLeaveRequest
      tags: [leave]
      summary: Delete leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get overall attendance summary
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Summary report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/employee/{userId}:
    get:
      operationId: getEmployeeReport
      tags: [reports]
      summary: Get attendance report for a specific employee
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Employee report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeReport"
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily attendance report
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyReportItem"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [superadmin, admin, employee]
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        shiftId:
          type: integer
          nullable: true
        shiftName:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userFullName:
          type: string
        departmentName:
          type: string
          nullable: true
        date:
          type: string
          format: date
        checkIn:
          type: string
          format: date-time
          nullable: true
        checkOut:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [present, absent, late, early_leave, on_leave]
        lateMinutes:
          type: integer
        earlyLeaveMinutes:
          type: integer
        workHours:
          type: number
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userFullName
        - date
        - status
        - lateMinutes
        - earlyLeaveMinutes
        - workHours
        - createdAt
...
    CheckInRequest:
      type: object
      properties:
        note:
          type: string
          nullable: true
    CheckOutRequest:
      type: object
      properties:
        note:
          type: string
          nullable: true
...
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userFullName:
          type: string
        leaveType:
          type: string
          enum: [sick, vacation, personal, other]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        days:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedById:
          type: integer
          nullable: true
        reviewedByName:
          type: string
          nullable: true
        reviewNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userFullName
        - leaveType
        - startDate
        - endDate
        - days
        - reason
        - status
        - createdAt
    CreateLeaveRequest:
      type: object
      properties:
        leaveType:
          type: string
          enum: [sick, vacation, personal, other]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
      required:
        - leaveType
        - startDate
        - endDate
        - reason
    UpdateLeaveRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reviewNote:
          type: string
          nullable: true
      required:
        - status
    ReportSummary:
      type: object
      properties:
        totalEmployees:
          type: integer
        presentCount:
          type: integer
        absentCount:
          type: integer
        lateCount:
          type: integer
        onLeaveCount:
          type: integer
        attendanceRate:
          type: number
        avgWorkHours:
          type: number
        byDepartment:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentSummary"
      required:
        - totalEmployees
        - presentCount
        - absentCount
        - lateCount
        - onLeaveCount
        - attendanceRate
        - avgWorkHours
        - byDepartment
    DepartmentSummary:
      type: object
      properties:
        departmentId:
          type: integer
        departmentName:
          type: string
        totalEmployees:
          type: integer
        presentCount:
          type: integer
        absentCount:
          type: integer
        attendanceRate:
          type: number
      required:
        - departmentId
        - departmentName
        - totalEmployees
        - presentCount
        - absentCount
        - attendanceRate
    EmployeeReport:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        totalDays:
          type: integer
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        onLeaveDays:
          type: integer
        totalWorkHours:
          type: number
        avgWorkHours:
          type: number
        attendanceRate:
          type: number
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
      required:
        - user
        - totalDays
        - presentDays
        - absentDays
        - lateDays
        - onLeaveDays
        - totalWorkHours
        - avgWorkHours
        - attendanceRate
        - records
    DailyReportItem:
      type: object
      properties:
        userId:
          type: integer
        userFullName:
          type: string
        departmentName:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        status:
          type: string
          enum: [present, absent, late, early_leave, on_leave]
        checkIn:
          type: string
          nullable: true
        checkOut:
          type: string
          nullable: true
        workHours:
          type: number
        lateMinutes:
          type: integer
      required:
        - userId
        - userFullName
        - status
        - workHours
        - lateMinutes
...